rename router.animations.ts and move it to classes
