+++ src/app/map/side-panel/side-panel.component.scss
@@ -17,5 +17,13 @@
     top: 45px;
     right: 10px;
   }
+
+  .tree-read-only {
+    width: 100%;
+  }
+
+  .tree-detail {
+    width: 100%;
+  }
 }
