+++ src/app/members/email/email.component.ts
@@ -1,22 +1,22 @@
-import {Component, EventEmitter, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, HostBinding, OnInit, Output} from '@angular/core';
 import {AuthService} from '../../services/auth.service';
 import { Router} from '@angular/router';
-import { moveIn, fallIn} from '../../router.animations';
+import { moveIn, fallIn} from '../../classes/animations';
 
 @Component({
   selector: 'app-email',
   templateUrl: './email.component.html',
   styleUrls: ['./email.component.scss'],
-  animations: [moveIn(),fallIn()],
-  host: {'[@moveIn]':''}
+  animations: [moveIn(), fallIn()],
 })
 export class EmailComponent implements OnInit {
-  state: string = '';
   error: any;
-  @Output() eventData:EventEmitter<string> = new EventEmitter();
+  @Output() eventData: EventEmitter<string> = new EventEmitter();
+  @HostBinding('@moveIn')
+  public state = true;
 
   constructor(private auth: AuthService, private router: Router) {
-    if(this.auth.isLoggedIn()){
+    if (this.auth.isLoggedIn()) {
       this.router.navigateByUrl('/members');
     }
   }
@@ -33,7 +33,7 @@ export class EmailComponent implements OnInit {
           }
         ).catch(
         (err) => {
-          this.error=err;
+          this.error = err;
         }
       );
     }
